requirements.txt
albumentations==1.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
docopt==0.6.2
idna==3.3
imageio==2.19.5
joblib==1.1.0
networkx==2.8.5
numpy==1.23.1
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.2.0
pipreqs==0.4.11
pyparsing==3.0.9
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.10
yarg==0.1.9
